Separate the handlers into hooks
JIRA: RAIL-4158
